requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-axes==6.5.1
django-cors-headers==4.4.0
django-formtools==2.5.1
django-otp==1.5.0
django-phonenumber-field==7.3.0
django-two-factor-auth==1.16.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
kombu==5.3.7
msgpack==1.0.8
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
PyJWT==2.8.0
pymongo==4.8.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
qrcode==7.4.2
redis==5.0.7
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
